requirements.txt
# Aegis Forensics Suite — Python Dependencies
# Install with: pip install -r requirements.txt

# ── GUI (Desktop App) ─────────────────────────────────────────────────────────
customtkinter>=5.2.2

# ── Metadata Extraction ───────────────────────────────────────────────────────
ExifRead>=3.0.0
PyMuPDF>=1.24.0

# ── Data Analysis (Log Parser) ────────────────────────────────────────────────
pandas>=2.2.0

# ── Web Application ───────────────────────────────────────────────────────────
flask>=3.0.0
Werkzeug>=3.0.0

# ── Network / API ─────────────────────────────────────────────────────────────
requests>=2.31.0

# ── Database ──────────────────────────────────────────────────────────────────
mysql-connector-python>=8.3.0
